fix: big reset
In this commit I'm reverting basically all of the CSS, styling
and component changes I've made since the beginning of this
project.
I'm not pleased with the way the design turned out and everything
seems really cluttered.
I'm remodeling everything, along with the components.
I'm gonna keep the features I've made in terms of backend and
API.