feature: added new components and styling
In this code session I created new components for the Home page.
The styling is still inline CSS and will soon be moved into a separate
file, considering that I need to use media queries for the resizing
of the SVG. Also, that makes the component looks cleaner and tidy.
Removed all of the old SVG's since they're no longer of use.