feature: account component WIP
I've tried passing the validation function inside the object literal,
but that didn't seem to work.
This is an interesting one, since I haven't found any references on the
web on this implementation, which makes me question its solidity.
One interesting thing I found out is that you can only call
a method or invoke a function inside an object literal if the object
itself has a named key.
This is one of the the reasons why I created the Field class and
implemented the FieldInterface interface (I know, very redundant.)
In case it's not very clear, my intention is to use the same component
for both forms of account validation: Sign up and Sign in.
In a desperate attempt of making this work I've tried passing an Yup
validation object as props to the validationSchema
property on the Form component created by Formik, but couldn't make it
work properly.
The only reference I found on the internet using these libraries was
with hard coded Fields.
Firmly believe there is something I'm missing, however, I'm starting
to lose hope in using dynamic validation for each field.
If I were to implement the simplest and most functional solution,
I would create two separate components, but that's no fun.

Again, this is a WIP, It's very messy and not the final implementation.